feat: auth login logout with integration tests Thorough integration tests required main.rs to be split into main.rs and app.rs for starting the app from tests.
