+++ music_browser/.env.example
@@ -4,7 +4,11 @@ RUST_LOG=info
 
 BACKUP_HOME="../"
 
-POCKETBASE_URL=https://127.0.0.1:8090
+# For local development, use HTTP. Switch to HTTPS in production with proper certs.
+POCKETBASE_URL=http://127.0.0.1:8090
 POCKETBASE_JWT_SECRET=change-me
+# Only needed for HTTPS with self-signed certs (production or advanced local setup)
+POCKETBASE_CA_CERT=
 AUTH_COOKIE_SECURE=true
 AUTH_ALLOW_EMPTY_JWT_SECRET=false
+AUTH_REQUIRE_LOGIN=false
