+++ music_browser/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod app;
 pub mod auth;
 pub mod db;
 pub mod jobs;
