+++ music_browser/templates/base.html
@@ -176,6 +176,16 @@
             <li><a href="/schedule">📅 Schedule</a></li>
             <li><a href="/profile">👤 Profile</a></li>
             <li class="nav-sep">│</li>
+            <li class="auth-nav" id="auth-nav-anon">
+                <a href="/login">🔐 Login</a>
+                <a href="/signup">✨ Sign Up</a>
+            </li>
+            <li class="auth-nav" id="auth-nav-authed" style="display:none;">
+                <form action="/logout" method="post" style="display:inline;">
+                    <button type="submit" style="background:none;border:none;color:var(--text);cursor:pointer;padding:0;font:inherit;">🚪 Logout</button>
+                </form>
+            </li>
+            <li class="nav-sep">│</li>
             <li class="nav-jobs"><a href="/jobs">🔧 Jobs</a></li>
         </ul>
     </nav>
@@ -203,6 +213,22 @@
             a.style.color = 'var(--accent)';
         }
     });
+})();
+
+// Auth nav toggle based on auth_present cookie
+(function() {
+    function getCookie(name) {
+        const value = '; ' + document.cookie;
+        const parts = value.split('; ' + name + '=');
+        if (parts.length === 2) return parts.pop().split(';').shift();
+        return null;
+    }
+    const anonNav = document.getElementById('auth-nav-anon');
+    const authedNav = document.getElementById('auth-nav-authed');
+    if (getCookie('auth_present')) {
+        if (anonNav) anonNav.style.display = 'none';
+        if (authedNav) authedNav.style.display = 'list-item';
+    }
 })();
 </script>
 </body>
