+++ music_browser/templates/login.html
@@ -30,6 +30,7 @@
                 <button type="submit" class="btn btn-primary">
                     Sign In
                 </button>
+                <a href="/signup" class="btn">Create an account</a>
             </div>
         </form>
 </div>
