+++ music_browser/templates/signup.html
@@ -1,0 +1,42 @@
+{% extends "base.html" %}
+
+{% block title %}Sign Up - Music Browser{% endblock %}
+
+{% block content %}
+<div class="page-header">
+    <h2>Create Account</h2>
+</div>
+
+<div class="form">
+        {% if let Some(err) = error_message %}
+        <div class="error">
+            {{ err }}
+        </div>
+        {% endif %}
+
+        <form action="/signup" method="post">
+            <div class="form-group">
+                <label for="email">Email</label>
+                <input type="email" name="email" id="email" autocomplete="email" required>
+            </div>
+
+            <div class="form-group">
+                <label for="password">Password</label>
+                <input type="password" name="password" id="password" autocomplete="new-password" minlength="12" required>
+                <p class="text-muted">Use at least 12 characters.</p>
+            </div>
+
+            <div class="form-group">
+                <label for="password_confirm">Confirm Password</label>
+                <input type="password" name="password_confirm" id="password_confirm" autocomplete="new-password" minlength="12" required>
+            </div>
+
+            <div class="form-actions">
+                <button type="submit" class="btn btn-primary">
+                    Create Account
+                </button>
+                <a href="/login" class="btn">Already have an account?</a>
+            </div>
+        </form>
+</div>
+{% endblock %}
